rename package domain to model
