--- core/src/main/java/br/com/jpcchaves/core/domain/Enum/UserRole.java
@@ -1,6 +1,0 @@
-package br.com.jpcchaves.core.domain.Enum;
-
-public enum UserRole {
-  USER,
-  ADMIN
-}
