+++ core/src/main/java/br/com/jpcchaves/core/model/Enum/UserRole.java
@@ -1,0 +1,6 @@
+package br.com.jpcchaves.core.model.Enum;
+
+public enum UserRole {
+  USER,
+  ADMIN
+}
