+++ core/src/main/java/br/com/jpcchaves/core/model/Role.java
@@ -1,4 +1,4 @@
-package br.com.jpcchaves.core.domain;
+package br.com.jpcchaves.core.model;
 
 public class Role {
   private Long id;
